m12n: move SpiHash to oak-commons
